RF_22.003.001 | Move to credentials.py
