+++ tests/test_credentials.py
@@ -1,6 +1,7 @@
 import pytest
 from selenium.common import NoSuchElementException
 from selenium.webdriver.common.by import By
+from selenium.webdriver.support.select import Select
 from selenium.webdriver.support.wait import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 
@@ -8,6 +9,7 @@ USERNAME = "Name"
 PASSWORD = "Password"
 DESCRIPTION = "Description: credential type is 'Username with password'"
 CREDENTIAL_ID = "1"
+
 
 def create_credential(driver, username, password, description, credential_id, treat_as_secret=False):
     driver.find_element(By.XPATH, "//*[@href = '/manage']").click()
@@ -49,6 +51,7 @@ def find_credential_card(driver, username=None, description=None, credential_id=
         )
         return driver.find_element(By.XPATH, card_found_by_username_and_description)
 
+
 def is_credential_present(driver, username=None, description=None, credential_id=None):
     try:
         if credential_id:
@@ -64,6 +67,7 @@ def is_credential_present(driver, username=None, description=None, credential_id
         return True
     except NoSuchElementException:
         return False
+
 
 @pytest.mark.dependency()
 def test_create(browser):
@@ -116,10 +120,12 @@ def test_delete(browser):
     empty_message = browser.find_element(By.XPATH, "//div[contains(text(), 'This credentials domain is empty')]")
 
     assert empty_message.is_displayed(), "Empty domain message not displayed after deletion"
-    assert not is_credential_present(browser, USERNAME, DESCRIPTION, CREDENTIAL_ID), "Credential still exists after deletion"
+    assert not is_credential_present(browser, USERNAME, DESCRIPTION,
+                                     CREDENTIAL_ID), "Credential still exists after deletion"
 
 
-@pytest.mark.parametrize("special_characters ",[
+@pytest.mark.skip(reason="ER_22.002.04")
+@pytest.mark.parametrize("special_characters ", [
     "!", "%", "&", "#", "@", "*"
 ])
 def test_create_id_with_special_characters(browser, special_characters):
@@ -136,3 +142,29 @@ def test_create_id_with_special_characters(browser, special_characters):
     )
 
     assert actual_error == expected_error
+
+
+def test_add_credentials_ssh_username(browser):
+    browser.find_element(By.ID, 'root-action-ManageJenkinsAction').click()
+    browser.find_element(By.CSS_SELECTOR, "a[href='credentials']").click()
+
+    browser.find_element(By.CSS_SELECTOR, "button[data-type='credentials-add-store-item']").click()
+    browser.find_element(By.XPATH, "//div[contains(text(),'SSH Username with private key')]").click()
+    browser.find_element(By.ID, 'cr-dialog-next').click()
+
+    Select(browser.find_element(By.NAME, '_.scope')).select_by_value('SYSTEM')
+    browser.find_element(By.NAME, '_.id').send_keys("testID")
+    browser.find_element(By.NAME, '_.description').send_keys("testDescription")
+    browser.find_element(By.NAME, '_.username').send_keys("testUsername")
+    browser.find_element(By.XPATH, "//label[normalize-space()='Treat username as secret']").click()
+    browser.find_element(By.XPATH, "//label[normalize-space()='Enter directly']").click()
+    browser.find_element(By.XPATH, "//button[normalize-space()='Add']").click()
+    browser.find_element(By.NAME, '_.privateKey').send_keys("testKey")
+    browser.find_element(By.NAME, '_.passphrase').send_keys("testPassphrase")
+    browser.find_element(By.ID, 'cr-dialog-submit').click()
+
+    (WebDriverWait(browser, 10)
+    .until(EC.presence_of_element_located(
+        (By.XPATH, "//*[contains(@class,'credentials-card') and contains(.,'testID')]"))))
+    cards = browser.find_elements(By.CSS_SELECTOR, ".credentials-card")
+    assert any("testID" in card.text for card in cards)
